requirements.txt
# FastAPI
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
asyncpg>=0.28.0
alembic>=1.12.0

# Redis
redis>=5.0.0

# Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# HTTP client
httpx>=0.24.0

# Payment
stripe>=6.0.0

# Push notifications
firebase-admin>=6.0.0

# Email
aiosmtplib>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0

# Utils
python-dotenv>=1.0.0
